back/services/user/requirements.txt
fastapi==0.116.1
uvicorn==0.32.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
email-validator==2.2.0